requirements.txt
fastapi
uvicorn
soundfile
webrtcvad
ffmpeg-python
numpy
python-multipart
